Refactored package structure and tests
