+++ README.md
@@ -2,12 +2,21 @@
 
 Collection of basic datastructures implemented in Java.
 
-* Dynamic Array
-* Hash Table
-* Linked List
+* Dynamic array
+* Hash table
+* Heap
+* Linked list
 * Queue
 * Stack
 * Tree
 * Trie
 
 Everything's public access for simpler testing.
+
+# Algorithms
+
+* Merge sort
+
+# Cracking The Code Interview
+
+Answers to the 6th Edition.
