+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>de.akull</groupId>
-    <artifactId>ds</artifactId>
+    <artifactId>algodat</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
