+++ src/main/java/de/akull/ctci/RemoveDups_2_1.java
@@ -1,7 +1,7 @@
 package de.akull.ctci;
 
-import de.akull.ds.linkedlist.LinkedList;
-import de.akull.ds.linkedlist.Node;
+import de.akull.datastructures.linkedlist.LinkedList;
+import de.akull.datastructures.linkedlist.Node;
 
 import java.util.ArrayList;
