+++ src/main/java/de/akull/datastructures/dynamicarray/DynamicArray.java
@@ -1,0 +1,20 @@
+package de.akull.datastructures.dynamicarray;
+
+import java.util.Arrays;
+
+@SuppressWarnings("WeakerAccess")
+public class DynamicArray<T> {
+
+    public static final int INCREASE_FACTOR = 2;
+
+    public int capacity = 3;
+    public int size = 0;
+    public Object[] array = new Object[capacity];
+
+    public void add(T t) {
+        if (size == capacity) {
+            array = Arrays.copyOf(array, capacity *= INCREASE_FACTOR);
+        }
+        array[size++] = t;
+    }
+}
