+++ src/main/java/de/akull/datastructures/hashtable/Node.java
@@ -1,0 +1,16 @@
+package de.akull.datastructures.hashtable;
+
+@SuppressWarnings("WeakerAccess")
+public class Node<K, V> {
+
+    public Node<K, V> next;
+
+    public K key;
+
+    public V value;
+
+    public Node(K key, V value) {
+        this.key = key;
+        this.value = value;
+    }
+}
