+++ src/main/java/de/akull/datastructures/linkedlist/Node.java
@@ -1,0 +1,12 @@
+package de.akull.datastructures.linkedlist;
+
+@SuppressWarnings("WeakerAccess")
+public class Node<T> {
+
+    public Node next;
+    public T data;
+
+    public Node(T data) {
+        this.data = data;
+    }
+}
