+++ src/main/java/de/akull/datastructures/tree/Node.java
@@ -1,0 +1,9 @@
+package de.akull.datastructures.tree;
+
+@SuppressWarnings("WeakerAccess")
+public class Node<T> {
+
+    public T data;
+    public Node left;
+    public Node right;
+}
