+++ src/main/java/de/akull/datastructures/trie/Trie.java
@@ -1,0 +1,29 @@
+package de.akull.datastructures.trie;
+
+@SuppressWarnings("WeakerAccess")
+public class Trie {
+    public Node root = new Node();
+
+    public void insert(String s) {
+        Node current = root;
+        for (char c : s.toCharArray()) {
+            current.children.putIfAbsent(c, new Node());
+            current = current.children.get(c);
+        }
+        current.isWord = true;
+    }
+
+    public boolean isWord(String s) {
+        Node current = root;
+
+        for (char c : s.toCharArray()) {
+            current = current.children.get(c);
+
+            if (current == null) {
+                return false;
+            }
+        }
+        return current.isWord;
+    }
+}
+
