--- src/main/java/de/akull/ds/hashtable/Node.java
@@ -1,16 +1,0 @@
-package de.akull.ds.hashtable;
-
-@SuppressWarnings("WeakerAccess")
-public class Node<K, V> {
-
-    public Node<K, V> next;
-
-    public K key;
-
-    public V value;
-
-    public Node(K key, V value) {
-        this.key = key;
-        this.value = value;
-    }
-}
