--- src/main/java/de/akull/ds/queue/Node.java
@@ -1,12 +1,0 @@
-package de.akull.ds.queue;
-
-@SuppressWarnings("WeakerAccess")
-public class Node<T> {
-
-    public T data;
-    public Node next;
-
-    public Node(T data) {
-        this.data = data;
-    }
-}
