--- src/main/java/de/akull/ds/stack/Node.java
@@ -1,12 +1,0 @@
-package de.akull.ds.stack;
-
-@SuppressWarnings("WeakerAccess")
-public class Node<T> {
-
-    public T data;
-    public Node next;
-
-    public Node(T data) {
-        this.data = data;
-    }
-}
