--- src/main/java/de/akull/ds/tree/Node.java
@@ -1,9 +1,0 @@
-package de.akull.ds.tree;
-
-@SuppressWarnings("WeakerAccess")
-public class Node<T> {
-
-    public T data;
-    public Node left;
-    public Node right;
-}
